fix senderId int에서 long으로 변경

+++ MoneyMoa_BE/src/main/java/com/d210/moneymoa/dto/ChatMessageDto.java
@@ -24,12 +24,12 @@ public class ChatMessageDto implements Serializable {
     private ChatMessage.MessageType type; // 메시지 타입
     private String roomId; // 방번호
     private String sender; // 메시지 보낸사람
-    private int senderId; // 메시지 보낸사람
+    private long senderId; // 메시지 보낸사람
     private String message; // 메시지
     private LocalDateTime createdTime; // 메시지 생성 시간
 
     @Builder
-    public ChatMessageDto(ChatMessage.MessageType type, String roomId, String sender, int senderId, String message) {
+    public ChatMessageDto(ChatMessage.MessageType type, String roomId, String sender, long senderId, String message) {
         this.type = type;
         this.roomId = roomId;
         this.sender = sender;
